+++ addon/route-handlers/function.js
@@ -1,6 +1,9 @@
-export default class FunctionRouteHandler {
+import BaseRouteHandler from './base';
+
+export default class FunctionRouteHandler extends BaseRouteHandler {
 
   constructor(schema, serializerOrRegistry, userFunction) {
+    super();
     this.schema = schema;
     this.serializerOrRegistry = serializerOrRegistry;
     this.userFunction = userFunction;
@@ -24,6 +27,12 @@ export default class FunctionRouteHandler {
     }
 
     return serializer.serialize(response, this.request);
+  }
+
+  normalizedRequestAttrs() {
+    let modelName = this.getModelClassFromPath(this.request.url);
+
+    return this._getAttrsForRequest(this.request, modelName);
   }
 
 }